Extract imagecontainer to outside of LoginPage
