+++ semana9/labex/src/pages/LoginPage/loginPage.styled.js
@@ -20,12 +20,3 @@ export const TitleLogin = styled.h2`
   font-size: 2em;
   margin-bottom: 10px;
 `;
-
-export const ImageContainer = styled.div`
-  width: 50%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
